chore(tests): port and re-enable `TestAdmissionWebhook_KongPlugins` and `TestAdmissionWebhook_KongClusterPlugins`
